requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.1
anyio==4.2.0
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
cohere==5.9.4
comm==0.1.4
contourpy==1.1.1
cycler==0.11.0
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.1.3
executing==1.2.0
fonttools==4.42.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.40
google-api-core==2.3.2
google-api-python-client==2.34.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-trace==1.5.1
googleapis-common-protos==1.54.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.15.0
jedi==0.19.0
jupyter_client==8.3.1
jupyter_core==5.3.1
kiwisolver==1.4.5
llamaapi==0.1.36
matplotlib==3.8.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.0
openai==1.30.1
packaging==23.1
pandas==2.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic_core==2.18.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.1
requests==2.31.0
seaborn==0.12.2
six==1.16.0
scikit-learn==1.3.2
smmap==5.0.1
sniffio==1.3.0
stack-data==0.6.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.4
wcwidth==0.2.6
yarl==1.9.2
zipp==3.17.0
